+++ packages/amplify-graphql-api-construct/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.9.0...@aws-amplify/graphql-api-construct@1.9.1) (2024-05-01)
+
+**Note:** Version bump only for package @aws-amplify/graphql-api-construct
+
 # [1.9.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.8.1...@aws-amplify/graphql-api-construct@1.9.0) (2024-04-26)
 
 ### Features