+++ packages/amplify-graphql-migration-tests/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.28](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-graphql-migration-tests@2.4.27...amplify-category-api-graphql-migration-tests@2.4.28) (2024-05-01)
+
+**Note:** Version bump only for package amplify-category-api-graphql-migration-tests
+
 ## [2.4.27](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-graphql-migration-tests@2.4.26...amplify-category-api-graphql-migration-tests@2.4.27) (2024-04-26)
 
 **Note:** Version bump only for package amplify-category-api-graphql-migration-tests